backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend of the sprite-splitter app."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.112.2",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "uvicorn>=0.30.6",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.3",
]
